COB: implemented StartBuilding
heading and pitch for this function are passed around as radians
in the engine and only in CCobInstance coverted to TA angles.

for this I needed to introduce GetHeadingFromVectorF and GetHAndPFromVectorF
(float versions of similarly named -but without F suffix- short int versions)

+++ rts/Sim/Units/COB/CobInstance.h
@@ -81,7 +81,7 @@ public:
 	virtual void TransportPickup(CUnit* unit);
 	virtual void EndTransport();
 	virtual void TransportDrop(CUnit* unit, const float3& pos);
-	virtual void StartBuilding(int heading, int pitch);
+	virtual void StartBuilding(float heading, float pitch);
 	virtual void StopBuilding();
 	virtual int  QueryNanoPiece();
